fix: restore standalone ipld instance
This PR partially  reverts
https://github.com/ipfs/ipld-explorer-components/pull/287
because `/api/v0/dag` in go-ipfs do not implement latest IPLD features
yet, and ipfs-http-client delegates path traversal and resolution to
go-ipfs, so codecs in JS are never used.

By reverting to standalone js-ipld we decouple IPLD explorer from what
is available in go-ipfs – we simply fetch raw block and do all decoding
the old way, in JS.

We will revisit this in the future, after IPLD Prime work lands in
go-ipfs, but for now this is the only way to fix IPLD Explorer to work
against go-ipfs backend.

License: MIT
Signed-off-by: Marcin Rataj <lidel@lidel.org>