fix: interop with modern JS API
This wrapper ensures the new block service
from js-ipfs AND js-ipfs-http-client works with the legacy code
present in ipld-explorer-components

(ir)rationale: we have no bandwidth to rewrite entire IPLD Explorer
..but thanks to it using only ipfs.block.get, making it extra compatible
is not very expensive. This buys us some time, but this technical debt
needs to be paid eventually.

License: MIT
Signed-off-by: Marcin Rataj <lidel@lidel.org>